requirements.txt
flask==3.1.2
langchain==0.2.16
langchain-openai==0.1.25
langchain-core==0.2.39
langchain-community==0.2.16
requests==2.32.3
python-dotenv==1.0.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
openai==1.35.14
